backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
celery>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
